fix ENV VARS
