--- server/.env
@@ -1,5 +1,0 @@
-db_host='192.168.32.3',
-db_port='5432',
-db_database='studhack',
-db_user='root',
-db_password='root'
